Refactor: myInfoForm 레이아웃 수정
