+++ src/components/myInfo/MyInfo.tsx
@@ -1,11 +1,5 @@
 import MyInfoForm from './myInfoForm';
-import ProfileModifyForm from './myInfoForm/profileModifyForm';
 
 export default function MyInfo() {
-  return (
-    <div>
-      <ProfileModifyForm />
-      <MyInfoForm />
-    </div>
-  );
+  return <MyInfoForm />
 }
